Refactor: Extract new mod `detect` for `detect-targets`
